Feat/rooms

+++ backend/src/main/java/com/casino/project/service/RoomService.java
@@ -4,9 +4,11 @@ import com.casino.project.dto.room.RoomCreateRequest;
 import com.casino.project.dto.room.RoomResponse;
 import com.casino.project.dto.room.RoomUpdateRequest;
 import com.casino.project.model.Room;
+import com.casino.project.model.RoomStatus;
 import com.casino.project.model.User;
 import com.casino.project.repository.RoomRepository;
 import com.casino.project.repository.UserRepository;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.data.domain.Page;
@@ -24,10 +26,12 @@ public class RoomService {
     }
 
     @Transactional
-    public RoomResponse createRoom(RoomCreateRequest request) {
+    public RoomResponse createRoom(RoomCreateRequest request, String username) {
         Room room = new Room();
+        User user = userRepository.findByUsername(username).orElseThrow(() -> new IllegalArgumentException("No player found with username " + username));
         room.setName(request.name());
         room.setCapacity(request.capacity());
+        room.setOwner(user);
         Room savedRoom = roomRepository.save(room);
         return mapToResponse(savedRoom);
     }
@@ -59,25 +63,83 @@ public class RoomService {
     }
 
     @Transactional
-    public RoomResponse updateRoom(Long roomId, RoomUpdateRequest request) {
-        Room room = roomRepository.findById(roomId).orElseThrow(() -> new RuntimeException("No room found with id " + roomId));
+    public RoomResponse updateRoom(Long roomId, RoomUpdateRequest request, String username) {
+        Room room = roomRepository.findById(roomId)
+                .orElseThrow(() -> new RuntimeException("No room found with id " + roomId));
+
+        if (!room.getOwner().getUsername().equals(username)) {
+            throw new AccessDeniedException("You are not the owner of this room");
+        }
         room.setName(request.name());
         room.setCapacity(request.capacity());
-        Room updatedRoom = roomRepository.save(room);
-        return mapToResponse(updatedRoom);
+        return mapToResponse(room);
     }
 
     @Transactional
-    public void deleteRoom(Long roomId) {
+    public void deleteRoom(Long roomId, String username) {
         Room room = roomRepository.findById(roomId).orElseThrow(() -> new RuntimeException("No room found with id " + roomId));
+        if (!room.getOwner().getUsername().equals(username)) {
+            throw new AccessDeniedException("You are not the owner of this room");
+        }
         roomRepository.delete(room);
     }
 
     @Transactional
-    public void joinRoom(Long userId, Long roomId) {
+    public RoomResponse joinRoom(String username, Long roomId) {
         Room room = roomRepository.findById(roomId).orElseThrow(() -> new RuntimeException("No room found with id " + roomId));
-        User user = userRepository.findById(userId).orElseThrow(() -> new IllegalArgumentException("No player found with id " + userId));
+        User user = userRepository.findByUsername(username).orElseThrow(() -> new IllegalArgumentException("No player found with id " + username));
 
         user.setCurrentRoom(room);
+        return mapToResponse(room);
+    }
+
+    @Transactional
+    public RoomResponse leaveRoom(String username, Long roomId) {
+        Room room = roomRepository.findById(roomId).orElseThrow(() -> new RuntimeException("No room found with id " + roomId));
+        User user = userRepository.findByUsername(username).orElseThrow(() -> new IllegalArgumentException("No player found with username " + username));
+        user.setCurrentRoom(null);
+        return mapToResponse(room);
+    }
+
+    @Transactional
+    public RoomResponse startGame(Long roomId, String username) {
+        Room room = roomRepository.findById(roomId).orElseThrow(() -> new RuntimeException("No room found with id " + roomId));
+
+        if (!room.getOwner().getUsername().equals(username)) {
+            throw new AccessDeniedException("Only owner can start the game");
+        }
+
+        if (room.getStatus() != RoomStatus.WAITING) {
+            throw new IllegalStateException("Game already in progress or finished");
+        }
+
+        if (room.getPlayers().size() < room.getCapacity()) {
+            throw new IllegalStateException("Not enough players to start");
+        }
+
+        room.setStatus(RoomStatus.IN_GAME);
+        // Game state
+
+        return mapToResponse(room);
+    }
+
+    @Transactional
+    public RoomResponse gameStatus(Long roomId, String username) {
+        Room room = roomRepository.findById(roomId).orElseThrow(() -> new RuntimeException("No room found with id " + roomId));
+        // Game state
+        return mapToResponse(room);
+    }
+
+    @Transactional
+    public RoomResponse gameAction(Long roomId, String username /* Action body here */) {
+        Room room = roomRepository.findById(roomId).orElseThrow(() -> new RuntimeException("No room found with id " + roomId));
+        // Game state
+        return mapToResponse(room);
+    }
+    @Transactional
+    public RoomResponse endGame(Long roomId, String username) {
+        Room room = roomRepository.findById(roomId).orElseThrow(() -> new RuntimeException("No room found with id " + roomId));
+        // Game state
+        return mapToResponse(room);
     }
 }
